Remove custom virtual column decorator
The custom virtual column decorator was removed from the codebase. Entity classes in User module were updated to leverage built-in TypeORM decorators, which resulted in cleaner and more maintainable code. The chance was made in response to a shift in project requirements that prioritizes conformity with established libraries and standards for maintainability. This change is not expected to introduce any differences in runtime behavior.